refactor(tests): split test_layer_dependencies.py into focused modules
Split the 1411 LOC test_layer_dependencies.py into smaller, more
maintainable test files:

- test_domain_purity.py: AST checks for domain layer (frozen dataclasses,
  I/O checks, complexity, Protocol usage)
- test_adapter_contracts.py: Infrastructure adapter tests (health_check,
  port compliance, atomic writes)
- test_forbidden_imports.py: Banned frameworks and import isolation
  (orchestration, metrics, facade imports)
- test_documentation.py: Docstring requirement checks for ports,
  adapters, and pipelines

Result: test_layer_dependencies.py reduced from 1411 to 761 LOC.
All 38 tests pass (with 3 skips for optional deps).

Refs: maintainability improvement task
