+++ tests/architecture/test_layer_dependencies.py
@@ -6,6 +6,10 @@ These tests verify that the clean architecture layer boundaries are respected:
 - Infrastructure layer: Implements Domain ports, can depend on external libraries
 
 Uses both static analysis and import-linter for comprehensive checks.
+
+Note: Tests for domain purity (frozen dataclasses, I/O checks, complexity) have been
+moved to test_domain_purity.py. Adapter contract tests moved to test_adapter_contracts.py.
+Forbidden import tests moved to test_forbidden_imports.py.
 """
 
 from __future__ import annotations
@@ -178,31 +182,6 @@ def test_ports_defined_in_domain_layer(src_dir: Path) -> None:
     assert protocol_found, "Ports should be defined using typing.Protocol"
 
 
-def test_infrastructure_imports_domain_ports(src_dir: Path) -> None:
-    """Infrastructure adapters should import from domain layer.
-
-    REQ-ARCH-006: Infrastructure implementations should implement domain ports.
-    """
-    adapters_path = src_dir / "bioetl" / "infrastructure" / "adapters"
-    if not adapters_path.exists():
-        pytest.skip("Infrastructure adapters not found")
-
-    found_domain_import = False
-    for py_file in adapters_path.rglob("*.py"):
-        if py_file.name == "__init__.py":
-            continue
-        with py_file.open(encoding="utf-8") as f:
-            content = f.read()
-        if "bioetl.domain" in content:
-            found_domain_import = True
-            break
-
-    assert found_domain_import, (
-        "Infrastructure adapters should import from domain layer "
-        "(e.g., to implement ports)"
-    )
-
-
 def test_import_linter_contracts(project_root: Path, src_dir: Path) -> None:
     """Run import-linter to verify all architectural contracts.
 
@@ -258,91 +237,13 @@ def test_infrastructure_does_not_import_application(src_dir: Path) -> None:
     forbidden = {"bioetl.application"}
 
     for py_file in infra_path.rglob("*.py"):
-        # ИСКЛЮЧЕНИЕ: config.py может импортировать PipelineConfig
+        # EXCEPTION: config.py can import PipelineConfig
         if py_file.name == "config.py":
             continue
         errors = _check_imports_in_file(py_file, forbidden)
         all_errors.extend(errors)
 
     assert not all_errors, "\n".join(all_errors)
-
-
-def test_domain_layer_uses_protocol_for_ports(src_dir: Path) -> None:
-    """Domain layer should use Protocol for defining ports.
-
-    REQ-ARCH-009: Ports should be defined using typing.Protocol
-    for structural subtyping (duck typing with type safety).
-    """
-    ports_dir = src_dir / "bioetl" / "domain" / "ports"
-    if not ports_dir.exists():
-        pytest.skip("ports/ package not found")
-
-    # Check at least one port file uses Protocol
-    protocol_import_found = False
-    protocol_class_found = False
-
-    for port_file in ports_dir.glob("*.py"):
-        if port_file.name == "__init__.py":
-            continue
-        with port_file.open(encoding="utf-8") as f:
-            content = f.read()
-            if "from typing" in content and "Protocol" in content:
-                protocol_import_found = True
-            if "class" in content and "(Protocol)" in content:
-                protocol_class_found = True
-            if protocol_import_found and protocol_class_found:
-                break
-
-    assert protocol_import_found, "Domain ports should use typing.Protocol for interface definitions"
-    assert protocol_class_found, "Port interfaces should be classes inheriting from Protocol"
-
-
-def test_cyclomatic_complexity_domain_layer(src_dir: Path) -> None:
-    """Domain layer functions should have low cyclomatic complexity.
-
-    REQ-ARCH-010: Domain logic should be simple and testable.
-    Maximum CC = 5 for domain layer functions.
-    """
-    try:
-        from radon.complexity import cc_visit
-    except ImportError:
-        pytest.skip("radon not installed")
-
-    domain_path = src_dir / "bioetl" / "domain"
-    if not domain_path.exists():
-        pytest.skip("Domain layer not found")
-
-    # Exemptions for specific functions (baseline)
-    exemptions = {
-        "__post_init__": 7,  # Dataclass post-init validation
-        "SchemaEvolutionError": 7,  # Exception with detailed field tracking
-    }
-
-    violations = []
-    max_cc = 5  # Strict threshold for domain layer
-
-    for py_file in domain_path.rglob("*.py"):
-        if py_file.name.startswith("__"):
-            continue
-
-        with py_file.open(encoding="utf-8") as f:
-            content = f.read()
-
-        try:
-            results = cc_visit(content)
-            for item in results:
-                func_max_cc = exemptions.get(item.name, max_cc)
-                if item.complexity > func_max_cc:
-                    violations.append(
-                        f"{py_file}:{item.lineno} - {item.name}() "
-                        f"has CC={item.complexity} (max={func_max_cc})"
-                    )
-        except SyntaxError:
-            continue
-
-    assert (
-        not violations
-    ), f"Domain layer has functions with CC > {max_cc}:\n" + "\n".join(violations)
 
 
 def test_no_empty_source_files(src_dir: Path) -> None:
@@ -515,32 +416,6 @@ def test_dead_code_vulture(src_dir: Path) -> None:
         )
 
 
-def test_application_layer_no_orchestration_imports(src_dir: Path) -> None:
-    """Application layer must not import orchestration frameworks directly.
-
-    REQ-ARCH-APP-001: Prefect, Celery, Airflow etc. должны быть изолированы
-    в отдельном слое (bioetl/interfaces/orchestration/).
-    """
-    application_path = src_dir / "bioetl" / "application"
-    if not application_path.exists():
-        pytest.skip("Application layer not found")
-
-    disallowed = ["prefect", "celery", "airflow", "dagster"]
-    violations = []
-
-    for py_file in application_path.rglob("*.py"):
-        content = py_file.read_text(encoding="utf-8")
-        for lib in disallowed:
-            if f"from {lib}" in content or f"import {lib}" in content:
-                violations.append(f"{py_file.relative_to(src_dir)}: imports {lib}")
-
-    assert not violations, (
-        "Application layer has direct orchestration imports:\n"
-        + "\n".join(f"  - {v}" for v in violations)
-        + "\n\nMove orchestration code to bioetl/interfaces/orchestration/"
-    )
-
-
 def test_application_layer_no_infrastructure_imports(src_dir: Path) -> None:
     """Application layer must not import from infrastructure.
 
@@ -593,66 +468,6 @@ def test_application_layer_no_infrastructure_imports(src_dir: Path) -> None:
     )
 
 
-def test_domain_value_objects_are_frozen(src_dir: Path) -> None:
-    """Domain Value Objects (dataclasses) must be frozen.
-
-    REQ-ARCH-014: Domain entities and value objects must be immutable
-    to ensure side-effect-free behavior and thread safety.
-    """
-    import ast
-
-    domain_path = src_dir / "bioetl" / "domain"
-    if not domain_path.exists():
-        pytest.skip("Domain layer not found")
-
-    violations = []
-
-    for py_file in domain_path.rglob("*.py"):
-        with py_file.open(encoding="utf-8") as f:
-            try:
-                tree = ast.parse(f.read(), filename=str(py_file))
-            except SyntaxError:
-                continue
-
-        for node in ast.walk(tree):
-            if isinstance(node, ast.ClassDef):
-                # Check for @dataclass decorator
-                is_dataclass = False
-                is_frozen = False
-
-                for decorator in node.decorator_list:
-                    # Case 1: @dataclass
-                    if isinstance(decorator, ast.Name) and decorator.id == "dataclass":
-                        is_dataclass = True
-                        # Default is frozen=False
-
-                    # Case 2: @dataclass(...)
-                    elif isinstance(decorator, ast.Call):
-                        func = decorator.func
-                        if isinstance(func, ast.Name) and func.id == "dataclass":
-                            is_dataclass = True
-                            # Check keywords for frozen=True
-                            for keyword in decorator.keywords:
-                                if (
-                                    keyword.arg == "frozen"
-                                    and isinstance(keyword.value, ast.Constant)
-                                    and keyword.value.value is True
-                                ):
-                                    is_frozen = True
-
-                if is_dataclass and not is_frozen:
-                    # Exemptions can be added here if strictly necessary, but default rule is strict
-                    violations.append(
-                        f"{py_file.name}:{node.lineno} - {node.name} is not frozen"
-                    )
-
-    assert (
-        not violations
-    ), "Found mutable domain dataclasses (must be frozen=True):\n" + "\n".join(
-        f"  - {v}" for v in violations
-    )
-
-
 def test_infrastructure_does_not_import_interfaces(src_dir: Path) -> None:
     """Infrastructure layer must not import from interfaces layer.
 
@@ -693,72 +508,6 @@ def test_infrastructure_does_not_import_composition(src_dir: Path) -> None:
 
     assert not all_errors, (
         "Infrastructure must not import composition.\n" + "\n".join(all_errors)
-    )
-
-
-def test_no_mutable_defaults_in_frozen_dataclasses(src_dir: Path) -> None:
-    """Frozen dataclasses should not have mutable default arguments.
-
-    REQ-ARCH-016: Mutable defaults (list, dict, set) in dataclasses
-    cause shared state issues even if the class is frozen.
-    """
-    import ast
-
-    bioetl_path = src_dir / "bioetl"
-    if not bioetl_path.exists():
-        pytest.skip("bioetl source not found")
-
-    violations = []
-
-    for py_file in bioetl_path.rglob("*.py"):
-        with py_file.open(encoding="utf-8") as f:
-            try:
-                tree = ast.parse(f.read(), filename=str(py_file))
-            except SyntaxError:
-                continue
-
-        for node in ast.walk(tree):
-            if isinstance(node, ast.ClassDef):
-                # Check if it's a dataclass
-                is_dataclass = False
-                for decorator in node.decorator_list:
-                    if (
-                        isinstance(decorator, ast.Name) and decorator.id == "dataclass"
-                    ) or (
-                        isinstance(decorator, ast.Call)
-                        and (
-                            isinstance(decorator.func, ast.Name)
-                            and decorator.func.id == "dataclass"
-                        )
-                    ):
-                        is_dataclass = True
-
-                if not is_dataclass:
-                    continue
-
-                # Check fields for mutable defaults
-                for item in node.body:
-                    if isinstance(item, ast.AnnAssign):
-                        if item.value:  # Has a default value
-                            is_mutable = False
-                            if isinstance(item.value, (ast.List, ast.Dict, ast.Set)):
-                                is_mutable = True
-                            elif isinstance(item.value, ast.Call):
-                                # Check for simple calls like list(), dict(), set()
-                                if isinstance(
-                                    item.value.func, ast.Name
-                                ) and item.value.func.id in ("list", "dict", "set"):
-                                    is_mutable = True
-
-                            if is_mutable:
-                                violations.append(
-                                    f"{py_file.name}:{item.lineno} - Field '{getattr(item.target, 'id', 'unknown')}' "
-                                    f"in class '{node.name}' has a mutable default value."
-                                )
-
-    assert not violations, (
-        "Found mutable defaults in dataclasses (use field(default_factory=...) instead):\n"
-        + "\n".join(f"  - {v}" for v in violations)
     )
 
 
@@ -845,203 +594,6 @@ def test_no_hasattr_duck_typing_in_application(src_dir: Path) -> None:
         "Found hasattr duck-typing in application layer. "
         "Add missing methods to port contracts in domain/ports/ package:\n"
         + "\n".join(f"  - {v}" for v in violations)
-    )
-
-
-def test_metrics_server_only_in_composition(src_dir: Path) -> None:
-    """Verify start_metrics_server is only called from composition layer.
-
-    REQ-ARCH-OBS-001: Observability initialization should only happen
-    in the composition root to ensure single point of responsibility.
-    """
-    forbidden_layers = ["interfaces", "application", "domain"]
-    allowed_patterns = [
-        r"def start_metrics_server",  # Definition is allowed
-        r"from.*import.*start_metrics_server",  # Import is allowed
-        r"#.*start_metrics_server",  # Comments are allowed
-        r"\"\"\".*start_metrics_server",  # Docstrings are allowed
-    ]
-
-    violations = []
-
-    for layer in forbidden_layers:
-        layer_path = src_dir / "bioetl" / layer
-        if not layer_path.exists():
-            continue
-
-        for py_file in layer_path.rglob("*.py"):
-            with py_file.open(encoding="utf-8") as f:
-                content = f.read()
-                lines = content.splitlines()
-
-            for i, line in enumerate(lines, 1):
-                # Check if line contains actual call to start_metrics_server
-                if "start_metrics_server(" in line:
-                    # Skip if matches allowed patterns
-                    if any(re.search(p, line) for p in allowed_patterns):
-                        continue
-
-                    relative_path = py_file.relative_to(src_dir)
-                    violations.append(f"{relative_path}:{i} - {line.strip()}")
-
-    assert not violations, (
-        "start_metrics_server() should only be called from composition layer.\n"
-        "Found calls in forbidden layers:\n"
-        + "\n".join(f"  - {v}" for v in violations)
-    )
-
-
-def test_no_direct_io_in_domain(src_dir: Path) -> None:
-    """Verify domain layer has no direct I/O operations.
-
-    REQ-ARCH-003: Domain layer should be pure business logic without I/O.
-    """
-    domain_path = src_dir / "bioetl" / "domain"
-    if not domain_path.exists():
-        pytest.skip("Domain layer not found")
-
-    # Patterns that indicate direct I/O
-    io_patterns = [
-        (r"\bopen\s*\(", "open() file access"),
-        (r"Path\s*\([^)]+\)\s*\.\s*(read|write|mkdir|unlink)", "Path I/O methods"),
-        (r"os\.(read|write|mkdir|remove|rename)", "os module I/O"),
-        (r"shutil\.(copy|move|rmtree)", "shutil I/O operations"),
-    ]
-
-    # Excluded files (test files, __init__.py)
-    excluded_files = {"__init__.py"}
-
-    violations = []
-
-    for py_file in domain_path.rglob("*.py"):
-        if py_file.name in excluded_files:
-            continue
-
-        with py_file.open(encoding="utf-8") as f:
-            content = f.read()
-            lines = content.splitlines()
-
-        for i, line in enumerate(lines, 1):
-            # Skip comments and docstrings
-            stripped = line.strip()
-            if stripped.startswith("#") or stripped.startswith('"""'):
-                continue
-
-            for pattern, description in io_patterns:
-                if re.search(pattern, line):
-                    relative_path = py_file.relative_to(src_dir)
-                    violations.append(
-                        f"{relative_path}:{i} - {description}: {stripped[:60]}..."
-                    )
-
-    assert not violations, (
-        "Domain layer should not have direct I/O operations.\n"
-        "Found violations:\n" + "\n".join(f"  - {v}" for v in violations)
-    )
-
-
-def test_atomic_write_used_in_writers(src_dir: Path) -> None:
-    """Verify that storage writers use atomic write patterns.
-
-    REQ-DATA-004: All file writes should be atomic to prevent data corruption.
-    """
-    storage_path = src_dir / "bioetl" / "infrastructure" / "storage"
-    if not storage_path.exists():
-        pytest.skip("Storage layer not found")
-
-    # Writers that should use atomic patterns
-    writer_files = ["bronze_writer.py", "gold_writer.py"]
-
-    # Patterns indicating non-atomic writes (reserved for future validation)
-    _non_atomic_patterns = [
-        r'with\s+open\s*\([^)]+,\s*["\']w',  # with open(path, 'w')
-        r'\.write\s*\([^)]+\)\s*$',  # .write() at end of line (might be in context)
-    ]
-
-    # Patterns indicating atomic writes (should be present)
-    atomic_indicators = [
-        r"atomic_write",
-        r"AtomicWriteGroup",
-        r"\.replace\s*\(",
-        r"tempfile\.mkstemp",
-    ]
-
-    findings = []
-
-    for writer_file in writer_files:
-        file_path = storage_path / writer_file
-        if not file_path.exists():
-            continue
-
-        with file_path.open(encoding="utf-8") as f:
-            content = f.read()
-
-        # Check for atomic indicators
-        has_atomic = any(re.search(p, content) for p in atomic_indicators)
-
-        if not has_atomic:
-            findings.append(f"{writer_file} - No atomic write patterns detected")
-
-    assert not findings, (
-        "Storage writers should use atomic write patterns (temp file + rename).\n"
-        "Files missing atomic patterns:\n" + "\n".join(f"  - {f}" for f in findings)
-    )
-
-
-def test_adapters_have_health_check(src_dir: Path) -> None:
-    """All adapters MUST implement health_check() method.
-
-    REQ-OBS-001: Adapters must provide health check for provider monitoring.
-    See docs/05-operations/runbooks/observability-checklist.md.
-    """
-    adapters_path = src_dir / "bioetl" / "infrastructure" / "adapters"
-    if not adapters_path.exists():
-        pytest.skip("Infrastructure adapters not found")
-
-    # Files that define adapter classes (not __init__.py or base classes)
-    # Exclude HTTP infrastructure utilities that are not DataSourcePort adapters
-    excluded_files = {
-        "base.py",
-        "types.py",
-        "exceptions.py",
-        "client.py",  # HTTP client utility, not a DataSourcePort adapter
-        "pagination.py",  # Pagination mixin, not a DataSourcePort adapter
-        "rate_limiter.py",  # Rate limiting utility
-        "circuit_breaker.py",  # Circuit breaker utility
-    }
-    adapter_files = []
-    for py_file in adapters_path.rglob("*.py"):
-        if py_file.name.startswith("_"):
-            continue
-        if py_file.name in excluded_files:
-            continue
-        adapter_files.append(py_file)
-
-    missing_health_check = []
-
-    for py_file in adapter_files:
-        content = py_file.read_text(encoding="utf-8")
-
-        # Check if file defines a class (likely an adapter)
-        if "class " not in content:
-            continue
-
-        # Check for health_check method definition
-        has_health_check = (
-            "def health_check" in content or "async def health_check" in content
-        )
-
-        if not has_health_check:
-            # Only flag if it looks like an adapter class
-            if "Adapter" in content or "Client" in content or "Fetcher" in content:
-                relative_path = py_file.relative_to(src_dir)
-                missing_health_check.append(str(relative_path))
-
-    assert not missing_health_check, (
-        "Adapters must implement health_check() method (REQ-OBS-001).\n"
-        "Files missing health_check:\n"
-        + "\n".join(f"  - {f}" for f in missing_health_check)
-        + "\n\nSee: docs/05-operations/runbooks/observability-checklist.md"
     )
 
 
@@ -1191,42 +743,6 @@ def test_storage_port_has_preview_cleanup(src_dir: Path) -> None:
     )
 
 
-def test_interfaces_no_direct_filesystem_traversal(src_dir: Path) -> None:
-    """Interfaces layer MUST NOT use direct filesystem traversal.
-
-    REQ-ARCH-023: CLI delegates to StoragePort, not Path.rglob.
-    """
-    interfaces_path = src_dir / "bioetl" / "interfaces"
-    if not interfaces_path.exists():
-        pytest.skip("Interfaces layer not found")
-
-    forbidden_patterns = [
-        r"\.rglob\(",
-        r"\.glob\(",
-        r"os\.walk\(",
-        r"os\.listdir\(",
-    ]
-
-    errors = []
-    for py_file in interfaces_path.rglob("*.py"):
-        if py_file.name.startswith("_"):
-            continue
-
-        with py_file.open(encoding="utf-8") as f:
-            content = f.read()
-
-        for pattern in forbidden_patterns:
-            if re.search(pattern, content):
-                relative_path = py_file.relative_to(src_dir)
-                errors.append(f"{relative_path}: contains '{pattern}'")
-
-    assert not errors, (
-        "Interfaces layer must not use direct filesystem traversal.\n"
-        "Delegate to StoragePort instead.\n"
-        "Violations:\n" + "\n".join(f"  - {e}" for e in errors)
-    )
-
-
 def test_error_classifier_uses_error_type_attribute(src_dir: Path) -> None:
     """ErrorClassifier SHOULD use error_type attribute for BioETLError.
 
@@ -1242,170 +758,4 @@ def test_error_classifier_uses_error_type_attribute(src_dir: Path) -> None:
     # Should have get_error_type() call for domain errors
     assert "get_error_type()" in content or "error_type" in content, (
         "ErrorClassifier should use error_type attribute for domain exceptions"
-    )
-
-
-def test_filterable_adapters_implement_protocol(src_dir: Path) -> None:
-    """Adapters with fetch_filtered MUST implement FilterableDataSourcePort.
-
-    REQ-ARCH-025: Replace duck-typing with explicit Protocol for adapters
-    that support filtering at API level. This ensures type safety and
-    enables isinstance() checks instead of hasattr().
-    """
-    adapters_path = src_dir / "bioetl" / "infrastructure" / "adapters"
-    if not adapters_path.exists():
-        pytest.skip("Infrastructure adapters not found")
-
-    violations = []
-
-    for py_file in adapters_path.rglob("*.py"):
-        if py_file.name.startswith("_"):
-            continue
-
-        content = py_file.read_text(encoding="utf-8")
-
-        # Check if file defines fetch_filtered method
-        has_fetch_filtered = (
-            "def fetch_filtered" in content or "async def fetch_filtered" in content
-        )
-
-        if has_fetch_filtered:
-            # Should reference FilterableDataSourcePort in docstring
-            has_protocol_ref = "FilterableDataSourcePort" in content
-
-            if not has_protocol_ref:
-                relative_path = py_file.relative_to(src_dir)
-                violations.append(
-                    f"{relative_path}: defines fetch_filtered but doesn't "
-                    "reference FilterableDataSourcePort"
-                )
-
-    assert not violations, (
-        "Adapters with fetch_filtered must implement FilterableDataSourcePort.\n"
-        "Update class/method docstrings to reference the Protocol:\n"
-        + "\n".join(f"  - {v}" for v in violations)
-    )
-
-
-def test_filtered_data_source_uses_isinstance(src_dir: Path) -> None:
-    """FilteredDataSource MUST use isinstance() for Protocol check.
-
-    REQ-ARCH-026: Replace hasattr() duck-typing with isinstance() check
-    for FilterableDataSourcePort. This enables proper type checking and
-    IDE support.
-    """
-    filtered_source = (
-        src_dir / "bioetl" / "application" / "core" / "filtered_data_source.py"
-    )
-    if not filtered_source.exists():
-        pytest.skip("FilteredDataSource not found")
-
-    content = filtered_source.read_text(encoding="utf-8")
-
-    # Should NOT use hasattr for fetch_filtered
-    uses_hasattr = 'hasattr' in content and 'fetch_filtered' in content
-    assert not uses_hasattr, (
-        "FilteredDataSource should not use hasattr() for fetch_filtered check. "
-        "Use isinstance(adapter, FilterableDataSourcePort) instead."
-    )
-
-    # Should use isinstance with FilterableDataSourcePort
-    uses_isinstance = (
-        "isinstance" in content and "FilterableDataSourcePort" in content
-    )
-    assert uses_isinstance, (
-        "FilteredDataSource must use isinstance(adapter, FilterableDataSourcePort) "
-        "for type-safe Protocol check."
-    )
-
-
-def test_bootstrap_no_direct_adapter_imports(src_dir: Path) -> None:
-    """bootstrap.py MUST NOT import concrete adapters directly.
-
-    REQ-ARCH-COMP-001: Composition Root delegates adapter creation to factories.
-    Adding a new provider should only require changes in:
-    - providers/registration.py (ProviderRegistry)
-    - factories/data_source_registry.py (DataSourceRegistry)
-
-    This prevents tight coupling and ensures the factory pattern is enforced.
-    """
-    bootstrap_file = src_dir / "bioetl" / "composition" / "bootstrap.py"
-    if not bootstrap_file.exists():
-        pytest.skip("bootstrap.py not found")
-
-    content = bootstrap_file.read_text(encoding="utf-8")
-
-    # Forbidden: direct imports of concrete adapter classes from provider packages
-    # Pattern: from bioetl.infrastructure.adapters.{provider}.{module} import {Class}
-    forbidden_patterns = [
-        (
-            r"from bioetl\.infrastructure\.adapters\.chembl\.\w+ import",
-            "ChEMBL adapter",
-        ),
-        (
-            r"from bioetl\.infrastructure\.adapters\.pubchem\.\w+ import",
-            "PubChem adapter",
-        ),
-        (
-            r"from bioetl\.infrastructure\.adapters\.uniprot\.\w+ import",
-            "UniProt adapter",
-        ),
-        (
-            r"from bioetl\.infrastructure\.adapters\.pubmed\.\w+ import",
-            "PubMed adapter",
-        ),
-    ]
-
-    violations = []
-    for pattern, description in forbidden_patterns:
-        matches = re.findall(pattern, content)
-        if matches:
-            violations.append(f"{description}: {matches}")
-
-    assert not violations, (
-        "bootstrap.py must not import concrete adapters directly.\n"
-        "Use ProviderRegistry and factories (DataSourceFactory, HttpClientFactory) instead.\n"
-        "Violations:\n" + "\n".join(f"  - {v}" for v in violations)
-    )
-
-
-def test_ports_imported_only_from_facade(src_dir: Path) -> None:
-    """Все слои MUST импортировать порты из фасада, не из внутренних модулей.
-
-    REQ-ARCH-027: Порты доступны только через bioetl.domain.ports,
-    а не через bioetl.domain.ports.storage и т.д.
-    Это обеспечивает единую точку входа и упрощает навигацию.
-    """
-    # Запрещённые паттерны импорта (внутренние модули пакета ports)
-    internal_port_modules = [
-        "bioetl.domain.ports.storage",
-        "bioetl.domain.ports.locking",
-        "bioetl.domain.ports.checkpoint",
-        "bioetl.domain.ports.quarantine",
-        "bioetl.domain.ports.observability",
-        "bioetl.domain.ports.data_source",
-        "bioetl.domain.ports.validation",
-        "bioetl.domain.ports.filtering",
-    ]
-
-    violations = []
-
-    # Проверить все слои кроме самого пакета ports
-    for layer in ["application", "composition", "infrastructure", "interfaces"]:
-        layer_path = src_dir / "bioetl" / layer
-        if not layer_path.exists():
-            continue
-
-        for py_file in layer_path.rglob("*.py"):
-            content = py_file.read_text(encoding="utf-8")
-            for module in internal_port_modules:
-                if f"from {module}" in content:
-                    relative_path = py_file.relative_to(src_dir)
-                    violations.append(f"{relative_path}: imports from {module}")
-
-    assert not violations, (
-        "Порты должны импортироваться только из фасада bioetl.domain.ports.\n"
-        "Правильно: from bioetl.domain.ports import StoragePort\n"
-        "Неправильно: from bioetl.domain.ports.storage import StoragePort\n\n"
-        "Нарушения:\n" + "\n".join(f"  - {v}" for v in violations)
     )
